+++ laravel/app/Http/Controllers/PermissionController.php
@@ -10,7 +10,8 @@ use App\Models\Role;
 
 class PermissionController extends Controller
 {
-    public function getRolePemissions(){
+    public function getRolePemissions(Request $request){
+        if(!$request->user()) return response()->json(['message'=>'Unauthenticated'],401);
         $permissions = PermessionVue_Role::with('permessionVue')->with('role')->get();
         $Data = [];
         foreach($permissions as $permission){
@@ -21,7 +22,8 @@ class PermissionController extends Controller
         }
         return response()->json(['permissions'=>$Data],200);
     } 
-    public function getRolePemissionsUsers(){
+    public function getRolePemissionsUsers(Request $request){
+        if(!$request->user()) return response()->json(['message'=>'Unauthenticated'],401);
         $permissions = permession_vues_users::with('permessionVue')->with('user')->get();
         $Data = [];
         foreach($permissions as $permission){
@@ -34,12 +36,14 @@ class PermissionController extends Controller
         return response()->json(['permissions'=>$Data]);
     } 
     public function ChangeStatusPermissionsUser(Request $request){
+        if(!$request->user()) return response()->json(['message'=>'Unauthenticated'],401);
         $permission = permession_vues_users::find($request->id);
         $permission->is_active = $request->is_active;
         $permission->save();
         return response()->json(['message'=>'Permission Updated'],200);
     }
-    public function getPemissionsAndRole(){
+    public function getPemissionsAndRole(Request $request){
+        if(!$request->user()) return response()->json(['message'=>'Unauthenticated'],401);
         $roles = Role::all();
         $rolesData = [];
         foreach($roles as $role){
@@ -61,6 +65,7 @@ class PermissionController extends Controller
     }
 
     public function addNewPermissions(Request $request){
+        if(!$request->user()) return response()->json(['message'=>'Unauthenticated'],401);
         foreach($request->formData['permissions_id'] as $data){
             $res = PermessionVue_Role::where('role_id',$request->formData["role_id"])->where('permession_vue_id',$data)->count();
             if($res == 0){
@@ -73,6 +78,7 @@ class PermissionController extends Controller
         return response()->json(['data'=>$request->formData],200);
     }
     public function deleteNewPermissions(Request $request){
+        if(!$request->user()) return response()->json(['message'=>'Unauthenticated'],401);
         PermessionVue_Role::where('id',$request->id)->delete();
         return response()->json(['data'=>$request->id],200);
     }
